+++ packages/jobs/lib/crons/autoIdleDemo.ts
@@ -1,5 +1,5 @@
 import { schedule } from 'node-cron';
-import { ErrorSourceEnum, LogActionEnum, SyncConfig, db, errorManager } from '@nangohq/shared';
+import { ErrorSourceEnum, SyncConfig, db, errorManager } from '@nangohq/shared';
 import tracer from 'dd-trace';
 
 export async function cronAutoIdleDemo(): Promise<void> {
@@ -9,7 +9,7 @@ export async function cronAutoIdleDemo(): Promise<void> {
             try {
                 await exec();
             } catch (err: unknown) {
-                const e = new Error('failed_to_clean_activity_logs_table', { cause: err instanceof Error ? err.message : err });
+                const e = new Error('failed_to_auto_idle_demo', { cause: err instanceof Error ? err.message : err });
                 errorManager.report(e, { source: ErrorSourceEnum.PLATFORM }, tracer);
             }
             span.finish();
