+++ packages/server/lib/server.ts
@@ -52,7 +52,7 @@ import oAuthSessionService from './services/oauth-session.service.js';
 import { deleteOldActivityLogs } from './jobs/index.js';
 import migrate from './utils/migrate.js';
 import tracer from './apm.js';
-import { cronAutoIdleDemo } from './jobs/cronAutoIdleDemo.js';
+import { cronAutoIdleDemo } from '../../jobs/lib/crons/autoIdleDemo.js';
 
 const { NANGO_MIGRATE_AT_START = 'true' } = process.env;
 
@@ -66,8 +66,8 @@ const webAuth =
     isCloud() || isEnterprise()
         ? [passport.authenticate('session'), authMiddleware.sessionAuth.bind(authMiddleware)]
         : isBasicAuthEnabled()
-        ? [passport.authenticate('basic', { session: false }), authMiddleware.basicAuth.bind(authMiddleware)]
-        : [authMiddleware.noAuth.bind(authMiddleware)];
+          ? [passport.authenticate('basic', { session: false }), authMiddleware.basicAuth.bind(authMiddleware)]
+          : [authMiddleware.noAuth.bind(authMiddleware)];
 
 app.use(
     express.json({
